--- frontend_v2/src/app/components/home/testimonials/testimonials.component.html
@@ -1,13 +1,0 @@
-<div class="testimonial-container" style="background-image: url('assets/images/testimonials/Testimonial.png');">
-  <div class="testimonial-body">
-    <span class="fs-24">{{ testimonialbody }}</span>
-  </div>
-  <div class="testimonial-author fs-18">- {{ testimonialauthor }}</div>
-  <img src="assets/images/testimonials/arrow.png" (click)="testimonialNavigate()" class="testimonial-arrow-left" />
-  <img
-    src="assets/images/testimonials/arrow.png"
-    (click)="testimonialNavigate('right')"
-    class="testimonial-arrow-right"
-  />
-  <img src="assets/images/testimonials/quotes.png" class="testimonial-quotes" />
-</div>
