--- frontend_v2/src/app/components/home/testimonials/testimonials.component.scss
@@ -1,71 +1,0 @@
-@import './variables.scss';
-@import './mixins.scss';
-
-.testimonial-container {
-  text-align: center;
-  background-size: 100% 100%;
-  background-repeat: no-repeat;
-  position: relative;
-  margin-top: 100px;
-  margin-bottom: 100px;
-  .testimonial-body {
-    width: 70%;
-    margin-left: 15%;
-    padding-top: 400px;
-    padding-bottom: 200px;
-    span {
-      color: #fff;
-      font-weight: $fw-light;
-      font-style: italic;
-      padding: 15px;
-    }
-  }
-
-  .testimonial-author {
-    color: #fff;
-    font-weight: $fw-regular;
-    padding-bottom: 200px;
-  }
-
-  .testimonial-arrow-left {
-    position: absolute;
-    top: 50%;
-    left: 0;
-    height: 100px;
-    cursor: pointer;
-  }
-
-  .testimonial-arrow-right {
-    position: absolute;
-    top: 50%;
-    right: 0;
-    height: 100px;
-    cursor: pointer;
-    transform: rotate(180deg);
-  }
-
-  .testimonial-quotes {
-    position: absolute;
-    top: 30%;
-    left: 10%;
-    height: 50px;
-  }
-}
-
-@include screen-medium {
-  .testimonial-container {
-    .testimonial-quotes {
-      top: 20% !important;
-      height: 40px !important;
-    }
-  }
-}
-
-@include screen-small {
-  .testimonial-container {
-    background-size: cover !important;
-    .testimonial-quotes {
-      height: 30px !important;
-    }
-  }
-}
